+++ src/main.cpp
@@ -5,15 +5,19 @@
 #include <sstream>
 
 // Parthenon headers
+#include "bvals/boundary_conditions_generic.hpp"
+#include "defs.hpp"
 #include "globals.hpp"
 #include "parthenon_manager.hpp"
 
 // AthenaPK headers
+#include "bvals/boundary_conditions_apk.hpp"
 #include "hydro/hydro.hpp"
 #include "hydro/hydro_driver.hpp"
 #include "main.hpp"
 
 #include "pgen/pgen.hpp"
+#include "tracers/tracers.hpp"
 // Initialize defaults for package specific callback functions
 namespace Hydro {
 InitPackageDataFun_t ProblemInitPackageData = nullptr;
@@ -23,6 +27,12 @@ SourceFun_t ProblemSourceUnsplit = nullptr;
 EstimateTimestepFun_t ProblemEstimateTimestep = nullptr;
 std::function<AmrTag(MeshBlockData<Real> *mbd)> ProblemCheckRefinementBlock = nullptr;
 } // namespace Hydro
+
+namespace Tracers {
+InitPackageDataFun_t ProblemInitTracerData = nullptr;
+SeedInitialFun_t ProblemSeedInitialTracers = nullptr;
+FillTracersFun_t ProblemFillTracers = nullptr;
+} // namespace Tracers
 
 int main(int argc, char *argv[]) {
   using parthenon::ParthenonManager;
@@ -55,6 +65,7 @@ int main(int argc, char *argv[]) {
     pman.app_input->InitUserMeshData = linear_wave_mhd::InitUserMeshData;
     pman.app_input->ProblemGenerator = linear_wave_mhd::ProblemGenerator;
     pman.app_input->UserWorkAfterLoop = linear_wave_mhd::UserWorkAfterLoop;
+    Hydro::ProblemInitPackageData = linear_wave_mhd::ProblemInitPackageData;
   } else if (problem == "cpaw") {
     pman.app_input->InitUserMeshData = cpaw::InitUserMeshData;
     pman.app_input->ProblemGenerator = cpaw::ProblemGenerator;
@@ -62,8 +73,8 @@ int main(int argc, char *argv[]) {
   } else if (problem == "cloud") {
     pman.app_input->InitUserMeshData = cloud::InitUserMeshData;
     pman.app_input->ProblemGenerator = cloud::ProblemGenerator;
-    pman.app_input->boundary_conditions[parthenon::BoundaryFace::inner_x2] =
-        cloud::InflowWindX2;
+    pman.app_input->RegisterBoundaryCondition(parthenon::BoundaryFace::inner_x2,
+                                              "cloud_inflow_x2", cloud::InflowWindX2);
     Hydro::ProblemCheckRefinementBlock = cloud::ProblemCheckRefinementBlock;
   } else if (problem == "blast") {
     pman.app_input->InitUserMeshData = blast::InitUserMeshData;
@@ -80,7 +91,9 @@ int main(int argc, char *argv[]) {
     pman.app_input->ProblemGenerator = field_loop::ProblemGenerator;
     Hydro::ProblemInitPackageData = field_loop::ProblemInitPackageData;
   } else if (problem == "kh") {
-    pman.app_input->ProblemGenerator = kh::ProblemGenerator;
+    pman.app_input->MeshProblemGenerator = kh::ProblemGenerator;
+  } else if (problem == "lw_implode") {
+    pman.app_input->ProblemGenerator = lw_implode::ProblemGenerator;
   } else if (problem == "rand_blast") {
     pman.app_input->ProblemGenerator = rand_blast::ProblemGenerator;
     Hydro::ProblemInitPackageData = rand_blast::ProblemInitPackageData;
@@ -97,6 +110,8 @@ int main(int argc, char *argv[]) {
   } else if (problem == "turbulence") {
     pman.app_input->MeshProblemGenerator = turbulence::ProblemGenerator;
     Hydro::ProblemInitPackageData = turbulence::ProblemInitPackageData;
+    Tracers::ProblemInitTracerData = turbulence::ProblemInitTracerData;
+    Tracers::ProblemFillTracers = turbulence::ProblemFillTracers;
     Hydro::ProblemSourceFirstOrder = turbulence::Driving;
     pman.app_input->InitMeshBlockUserData = turbulence::SetPhases;
     pman.app_input->MeshBlockUserWorkBeforeOutput = turbulence::UserWorkBeforeOutput;
@@ -106,6 +121,23 @@ int main(int argc, char *argv[]) {
     msg << "Problem ID '" << problem << "' is not implemented yet.";
     PARTHENON_THROW(msg);
   }
+
+  const std::string REFLECTING = "reflecting";
+  using BF = parthenon::BoundaryFace;
+  using Hydro::BoundaryFunction::ReflectBC;
+  using parthenon::BoundaryFunction::BCSide;
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x1, REFLECTING,
+                                            ReflectBC<X1DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x1, REFLECTING,
+                                            ReflectBC<X1DIR, BCSide::Outer>);
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x2, REFLECTING,
+                                            ReflectBC<X2DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x2, REFLECTING,
+                                            ReflectBC<X2DIR, BCSide::Outer>);
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x3, REFLECTING,
+                                            ReflectBC<X3DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x3, REFLECTING,
+                                            ReflectBC<X3DIR, BCSide::Outer>);
 
   pman.ParthenonInitPackagesAndMesh();